requirements.txt
numpy~=1.25.2
pandas~=2.1.1
plotly~=5.17.0
Pillow~=10.0.1
dash~=2.13.0
keras~=2.14.0
matplotlib~=3.8.0
scikit-learn~=1.3.1
tqdm~=4.66.1
tifffile~=2023.9.26
scikit-image~=0.21.0
opencv-python~=4.8.1.78
ray~=2.7.0
scipy~=1.11.3
joblib~=1.3.2

# additionally use the following commands:
# pip install git+https://github.com/zephyris/tryptag
# pip install scikit-learn-intelex
# pip install umap-learn